pyproject.toml
[project]
name = "brand-generator"
version = "0.2.0"
description = "Intelligent neologism generator for brand names with multi-provider LLM support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# Individual LLM providers
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.50.0"]
gemini = ["google-genai>=1.0.0"]

# All cloud providers
cloud = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-genai>=1.0.0",
]

# All providers (cloud + local support via httpx)
all = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-genai>=1.0.0",
]

dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[project.scripts]
brand-gen = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
